2. Conversational AI Chatbot/chatbot_train.py
from chatterbot import ChatBot
from chatterbot.trainers import ListTrainer
import os

def setup():
    chatbot = ChatBot('Bot',
    storage_adapter='chatterbot.storage.SQLStorageAdapter',
    trainer='chatterbot.trainers.ListTrainer')
    for file in os.listdir('C:/Users/amira/Desktop/Final Year Project/Application/data/'):
        convData = open(r'C:/Users/amira/Desktop/Final Year Project/Application/data/' +file, 'r').readlines()
        chatbot.set_trainer(ListTrainer)
        chatbot.train(convData)
        #print("Training completed")
    

setup()
