2. Conversational AI Chatbot/requirement.txt
tensorflow 2.0
keras
pandas
numpy
matplotlib
simple websocket server
ChatterBot-0.8.4

